perf(ortho-core): 頂点色をfloat32→Uint8正規化＝geometryバイト44%減
頂点色 RGBA を float32×4(16B) → Uint8×4(4B・normalized attrib) へ。fill 頂点は 24B→12B、
line インスタンスは 33B→21B。style の色は元々8bit精度（#hex/rgb）＝float32はバス幅の無駄だった。
tess出力 → postMessage transfer → scene worker 常駐 → merge memcpy → GL upload の全段が同時に縮み、
バイト予算LRU（45a8edb）の実質常駐枚数も約1.8倍になる。

- build: parseRGBA の float を feature 毎に一度だけ 0..255 へ丸め、col を Uint8Array で出力
- scene/tilemanager: merge の col 結合先を Uint8Array に（.set コピーは同型なので無変更）
- renderer: attrib() に u8 引数＝UNSIGNED_BYTE 正規化。シェーダは同じ 0..1 を受ける＝無変更。
  型判定（col instanceof Uint8Array）で geojson 由来の float32 シーンとは共存

検証: HEAD原版と実タイル37枚(z8-16 東京/渋谷/八王子/富士/横浜)で pos/P1/P2/half 全要素bit一致、
色は四捨五入の理論上限 0.5/255 以内（merge結合後も同様）。実測 geometry 16.28MB→9.20MB(56.5%)。

+++ packages/ortho-core/src/scene.js
@@ -19,8 +19,8 @@ export function mergeTiles(order, geomOf, opts = {}) {
 	const buf = new Map();
 	for (const [li, e] of size) {
 		buf.set(li, e.kind === "fill"
-			? { kind: "fill", li, pos: new Float32Array(e.fillN * 2), col: new Float32Array(e.fillN * 4), pi: 0, ci: 0 }
-			: { kind: "line", li, P1: new Float32Array(e.lineN * 2), P2: new Float32Array(e.lineN * 2), col: new Float32Array(e.lineN * 4), half: new Float32Array(e.lineN), pi: 0, ci: 0, hi: 0 });
+			? { kind: "fill", li, pos: new Float32Array(e.fillN * 2), col: new Uint8Array(e.fillN * 4), pi: 0, ci: 0 }
+			: { kind: "line", li, P1: new Float32Array(e.lineN * 2), P2: new Float32Array(e.lineN * 2), col: new Uint8Array(e.lineN * 4), half: new Float32Array(e.lineN), pi: 0, ci: 0, hi: 0 });
 	}
 	for (const { ox, oy, ops } of tileOps) {
 		for (const op of ops) {
